Change return type of linspaced_array

+++ stan/math/prim/fun/linspaced_array.hpp
@@ -23,7 +23,7 @@ namespace math {
  * @throw std::domain_error if K is negative, if low is nan or infinite,
  * if high is nan or infinite, or if high is less than low.
  */
-inline std::vector<int> linspaced_array(int K, double low, double high) {
+inline std::vector<double> linspaced_array(int K, double low, double high) {
   static const char* function = "linspaced_array";
   check_nonnegative(function, "size", K);
   check_finite(function, "low", low);
